Prove architecture
Using
[electron-vue](https://github.com/SimulatedGREG/electron-vue),
and some additions, this commit aims to prove that this architecture
will serve the purpose of editing the an-other-place game database.

The old mof editor was:

- Slow
- Ugly
- Awkward to run

This one uses VueJS instead of Angular1.x, which I hope will address the
speed issue. It uses vue-material, which should address the ugly issue.
And it uses electron, which means that it no longer requires a command
line application to run in the background and a browser to actually do
the editing.

I have hopes that:

- vuex will keep my architecture clean and maintainable
- separating out a quad editor from e.g. an impassable editor will help
with speed and code simplicity
- Aditional editing tools can be added to this - like a drop table
editor