Refactor Texture registered notification
This notification now visually coalesces multiple texture registrations
into the one message. Much happier with this behaviour.

I think one of the things that I'm unhappy with in the mof-editor was
too many cut corners. I was trying to be pragmatic about dev time vs
feature polish, and I think I got it wrong in some pretty fundamental
ways. The editor itself is actually difficult to use to place Quads, and
if it's on thing that it should be good at it's that.

I came back and fixed this UI display bug immediately, because I don't
want to repeat the mistakes of the past, so I'm trying for more polish
upfront.