Add a Level renderer
This was a very exploratory set of changes that I'm cutting here. It
introduces a couple of new pages, and modifies some existing ones. I
haven't split these into separate commits, because even though it's a
little unclear, I've been wrangling this for so long I don't remember
where I was up to.

This shows a list of levels, and shows all the layers of a level when
you open it.