WorldPainter: Add a selection tool
The user needs to be able to select existing tiles in the level in order
to move them, or to delete them.

I've moved away from the previous architecture I've been using to
communicate between World and WorldPainter here, because I've found what
I was doing unwieldy, verbose, and that the separation I gained from
doing it was not valuable. WorldPainter is tightly coupled to World, and
it could not be otherwise. I'm happy for it to call methods on World
directly.