World: Remove "cursor"
This was a concept that felt like a poor fit for the World component. It
was complicating the component, and felt like it was not necessary for
this component to implement it.

I've continued with the direct calls between WorldPainter and World, and
introduced a method on World to create and register tiles so that
clients of World can implement their own cursors. This might come in
handly if I ever want to be able to Copy & Paste multiple tiles, and
have the cursor be represented by the multiple tiles that can be Pasted.