World: Refactor keyboard events into new component
World doesn't need to own responsibility for listening to keyboard
events. This reusable component seems like a good way for components to
declare their interest in keyboard events.

Slightly concerned that this might cause issues if two different clients
of the KeyboardEvent component both listen to the same key event, but
this is probably a "cross that bridge when we come to it" issue.